+++ NEWS.md
@@ -12,6 +12,8 @@ NEW FEATURES:
 
  * Add save_noaa() to save a weather forecast to an RDS file.
 
+ * Add update_noaa() to save weather forcast to database if updated.
+
  * timezone(new) returns current time zone invisibly if a new one is set.
  
  * Argument 'days' for ggplot_noaa_{wind_direction,surface_wind}() can now